scripts/release/requirements.txt
poetry==1.1.4
gitpython==3.1.11
